+++ default.nix
@@ -2,8 +2,8 @@
 
 haskellPackages.cabal.mkDerivation (self: rec {
   pname = "libjenkins";
-  version = "0.6.0";
-  src = ./.;
+  version = "0.8.0";
+  src = builtins.filterSource (path: type: type != "unknown") ./.;
   buildDepends = with haskellPackages; [
     attoparsec free httpClient httpClientTls httpTypes liftedAsync
     liftedBase monadControl mtl network networkUri profunctors text