+++ test/Jenkins/RestSpec.hs
@@ -44,35 +44,9 @@ spec = do
       [QGet 0 "foo", QGet 1 "bar", QGet 2 "baz"]
 
 
-  describe "reload" $
-    it "calls $jenkins_url/reload with POST query and then disconnects" $ do
-      interpret $ do
-        Jenkins.reload
-        Jenkins.post_ "foo"
-     `shouldBe`
-      [QPost 0 "" "reload", QDisconnect]
-
-  describe "restart" $
-    it "calls $jenkins_url/safeRestart with POST query and then disconnects" $ do
-      interpret $ do
-        Jenkins.restart
-        Jenkins.post_ "bar"
-     `shouldBe`
-      [QPost 0 "" "safeRestart", QDisconnect]
-
-  describe "forceRestart" $
-    it "calls $jenkins_url/restart with POST query and then disconnects" $ do
-      interpret $ do
-        Jenkins.forceRestart
-        Jenkins.post_ "baz"
-     `shouldBe`
-      [QPost 0 "" "restart", QDisconnect]
-
-
 data Query =
     QGet Int Strict.ByteString
   | QPost Int Lazy.ByteString Strict.ByteString
-  | QDisconnect
     deriving (Show, Eq)
 
 newtype Requests a = Requests [a]
@@ -87,8 +61,6 @@ interpret adt = evalState (iter go ([] <$ adt)) (Requests [0..]) where
   go (Post m body n) = do
     r <- render (\x y -> QPost x body y) m
     fmap (r :) (n mempty)
-  go Dcon =
-    return [QDisconnect]
 
 render :: (a -> Strict.ByteString -> Query) -> Jenkins.Method Method.Complete f -> State (Requests a) Query
 render f m = do